🗃️ Change rating type in db to double

+++ lib/data/repositories/review_repository.dart
@@ -60,7 +60,7 @@ class ReviewRepository {
 
   Future<int> create({
     required String comment,
-    required int rating,
+    required double rating,
     required int itemId,
   }) {
     return _reviewsDao.create(comment: comment, rating: rating, itemId: itemId);
@@ -69,7 +69,7 @@ class ReviewRepository {
   Future<bool> updateById(
     int id, {
     required String comment,
-    required int rating,
+    required double rating,
     required int itemId,
   }) {
     return _reviewsDao.updateById(
