+++ lib/database/daos/reviews_dao.dart
@@ -35,7 +35,7 @@ class ReviewsDao extends DatabaseAccessor<AppDatabase> with _$ReviewsDaoMixin {
 
   Future<int> create({
     required String comment,
-    required int rating,
+    required double rating,
     required int itemId,
   }) {
     return into(reviews).insert(
@@ -50,7 +50,7 @@ class ReviewsDao extends DatabaseAccessor<AppDatabase> with _$ReviewsDaoMixin {
   Future<bool> updateById(
     int id, {
     required String comment,
-    required int rating,
+    required double rating,
     required int itemId,
   }) {
     return (update(reviews)..where((t) => t.id.equals(id)))
